Feature cellview

+++ PipeDream/PDGridView.h
@@ -13,7 +13,7 @@
 @interface PDGridView : UIView
 
 @property (nonatomic, weak) id<PDCellPressedDelegate> delegate;
-- (void) drawGridFromDimension: (int) gridDimension;
+- (void) drawGridFromDimension: (NSInteger) gridDimension;
 - (void) rotateClockwiseCellAtRow:(NSInteger)row col:(NSInteger)col;
 - (void) setCellAtRow:(NSInteger)row col:(NSInteger)col
           isOpenNorth:(BOOL)north east:(BOOL)east south:(BOOL)south west:(BOOL)west;
